Style : /:userId 라우터 경로 추가, loginId와 비교 로직 추가

+++ back/src/routers/userRouter.js
@@ -43,33 +43,38 @@ userAuthRouter.post("/login", async (req, res, next) => {
   }
 })
 
-userAuthRouter.get("/current", loginRequired, async (req, res, next) => {
+userAuthRouter.get("/:userId", loginRequired, async (req, res, next) => {
   try {
     // jwt토큰에서 추출된 사용자 id를 가지고 db에서 사용자 정보를 찾음.
-    const userId = req.currentUserId
-    const currentUserInfo = await userAuthService.getUserInfo({
-      userId,
-    })
+    const loginId = req.currentUserId
+    const userId = req.params.userId
+    if (loginId === userId) {
+      const currentUserInfo = await userAuthService.getUserInfo({
+        userId,
+      })
 
-    res.status(200).send(currentUserInfo)
+      res.status(200).send(currentUserInfo)
+    }
   } catch (error) {
     next(error)
   }
 })
 
-userAuthRouter.put("/current", loginRequired, async (req, res, next) => {
+userAuthRouter.put("/:userId", loginRequired, async (req, res, next) => {
   try {
     // URI로부터 사용자 id를 추출함.
-    const userId = req.currentUserId
-    // body data 로부터 업데이트할 사용자 정보를 추출함.
-    const { nickname, email, password, bookmarks } = req.body ?? null
+    const loginId = req.currentUserId
+    const userId = req.params.userId
+    if (loginId === userId) {
+      // body data로부터 업데이트할 사용자 정보를 추출함.
+      const { nickname, email, password, bookmarks } = req.body ?? null
+      const updateData = { nickname, email, password, bookmarks }
 
-    const updateData = { nickname, email, password, bookmarks }
+      // 해당 사용자 아이디로 사용자 정보를 db에서 찾아 업데이트함. 업데이트 요소가 없을 시 생략함
+      const updatedUser = await userAuthService.setUser({ userId, updateData })
 
-    // 해당 사용자 아이디로 사용자 정보를 db에서 찾아 업데이트함. 업데이트 요소가 없을 시 생략함
-    const updatedUser = await userAuthService.setUser({ userId, updateData })
-
-    res.status(200).json(updatedUser)
+      res.status(200).json(updatedUser)
+    }
   } catch (error) {
     next(error)
   }
@@ -77,26 +82,35 @@ userAuthRouter.put("/current", loginRequired, async (req, res, next) => {
 
 userAuthRouter.delete("/:userId", loginRequired, async (req, res, next) => {
   try {
+    const loginId = req.currentUserId
     const userId = req.params.userId
-    const result = await userAuthService.deleteUser({ userId })
+    if (loginId === userId) {
+      const result = await userAuthService.deleteUser({ userId })
 
-    res.status(200).send(result)
+      res.status(200).send(result)
+    }
   } catch (error) {
     next(error)
   }
 })
 // 게임 북마크/북마크 취소
 userAuthRouter.put(
-  "/bookmark/:gameId",
+  "/:userId/addBookmark",
   loginRequired,
   async (req, res, next) => {
     try {
-      const userId = req.currentUserId
-      const gameId = req.params.gameId
+      const loginId = req.currentUserId
+      const userId = req.params.userId
+      if (loginId === userId) {
+        const gameId = req.body.gameId
 
-      const updatedUser = await userAuthService.setBookmark({ userId, gameId })
+        const updatedUser = await userAuthService.setBookmark({
+          userId,
+          gameId,
+        })
 
-      res.status(200).send(updatedUser)
+        res.status(200).send(updatedUser)
+      }
     } catch (error) {
       next(error)
     }
@@ -104,16 +118,23 @@ userAuthRouter.put(
 )
 
 // 사용자별 북마크 리스트
-userAuthRouter.get("/bookmarks", loginRequired, async (req, res, next) => {
-  try {
-    const userId = req.currentUserId
-    const bookmarkList = await userAuthService.getBookmarkList({ userId })
+userAuthRouter.get(
+  "/:userId/bookmarks",
+  loginRequired,
+  async (req, res, next) => {
+    try {
+      const loginId = req.currentUserId
+      const userId = req.params.userId
+      if (loginId === userId) {
+        const bookmarkList = await userAuthService.getBookmarkList({ userId })
 
-    res.status(200).send(bookmarkList)
-  } catch (error) {
-    next(error)
+        res.status(200).send(bookmarkList)
+      }
+    } catch (error) {
+      next(error)
+    }
   }
-})
+)
 
 // jwt 토큰 기능 확인용, 삭제해도 되는 라우터임.
 userAuthRouter.get("/afterlogin", loginRequired, (req, res, next) => {
